Cleanup of code, moved everything to /src
