+++ .gitignore
@@ -1,0 +1,2 @@
+/SNMAPI.sprx
+/ppu/
