--- make_sprx.bat
@@ -1,10 +1,0 @@
-@echo off
-set PS3SDK=/c/PSDK3v2
-set WIN_PS3SDK=C:/PSDK3v2
-set PATH=%WIN_PS3SDK%/mingw/msys/1.0/bin;%WIN_PS3SDK%/mingw/bin;%WIN_PS3SDK%/ps3dev/bin;%WIN_PS3SDK%/ps3dev/ppu/bin;%WIN_PS3SDK%/ps3dev/spu/bin;%WIN_PS3SDK%/mingw/Python27;%PATH%;
-set PSL1GHT=%PS3SDK%/psl1ght
-set PS3DEV=%PS3SDK%/ps3dev
-set CELL_SDK=/c/usr/local/cell
-make
-
-pause
