+++ src/SNMAPI_Lib.h
