+++ src/assembly.h
