+++ src/helper_functions.h
