+++ src/libc.c
