+++ src/vsh_functions.h
